Include payedAttendee script on event pages
Change the include setting for the payedAttendee script so it is actually
included on event pages. Required $this->event to check for the presence
of an event, and $this->attendees to stop it from being included on the
"endre" page as well. These variables are attached to $this in the
index controller in arrangementbyidAction.