Make config file read only
When the config file was split in to, index.php was changed to load both
config files and merge them. In order to merge them the config being
merged into was made writable, and this writable config was then passed
to the application after merging.

This config is now again made read only before being passed to the
application, preserving the original behaviour.